Fixed use of landmarks

+++ geotnf/point_tnf.py
@@ -29,7 +29,7 @@ class PointTnf(object):
     
     def affPointTnf(self,theta,points):
         theta_mat = theta.view(-1,2,3)
-        warped_points = torch.bmm(theta_mat[:,:,:2],points)
+        warped_points = torch.bmm(theta_mat[:,:,:2].float(),points.float())
         warped_points += theta_mat[:,:,2].unsqueeze(2).expand_as(warped_points)
         return warped_points
